requirements.txt
PyQt5==5.9.2
fbs
pyyaml
PyInstaller==3.4
